transform test_mobexec_engine in Command
test_mobexec_engine inherit from Command instead of build

run remove the test to prevent if mobyle.common is installed
because when the module 'mobyle' is load is not reload.
even if the path is modified so the mobyle.exec_engine library are nver
found.
now the tests are run the results of tests are parsed and if the
mobyle.lib is not installed an explicit message is displayed.

the setup.py test now exit with a returncode = 1 if some tests failed